requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion Models
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0

# Image Processing
opencv-python>=4.7.0
scikit-image>=0.19.0
Pillow>=9.0.0

# Metrics & Evaluation
lpips>=0.1.4
scipy>=1.9.0
numpy>=1.21.0

# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# Database & Storage
sqlalchemy>=2.0.0

# Additional ML Libraries
timm>=0.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.0
requests>=2.28.0

# Optional: Experiment Tracking
# tensorboard>=2.11.0
# wandb>=0.13.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
